+ остатки для Goods, + при сохраннии документа поступления прибавляются остатки

+++ Documents/classes/TablePartString.cs
@@ -8,7 +8,7 @@ namespace Documents
 {
     public class TablePartString
     {
-        public byte serial;
+        public int serial;
         public Good Good;
         public int quantity;
         public int balance;
